+++ src/task2/api/get.js
@@ -1,0 +1,7 @@
+export function get(req, res, database) {
+    return res.status(200).send({
+        success: 'true',
+        message: 'persons retrieved successfully',
+        database
+    });
+}
